% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTrain.R
\name{genTrain}
\alias{genTrain}
\title{Generate auditory presentation}
\usage{
genTrain(sound, species = NULL, infodb = NULL, tpres, intRend.type,
  intRend.length = NA, int.factor = 1, nbout = NA, bout.length = NA,
  intBout.type, intBout.length = NA, rate = NA, seed = NA, writewav = T,
  output, ext = "wav", report = T, returnoutput = F)
}
\arguments{
\item{sound}{Source sound. Can also be a directory, or many directories,
containing the sound file.}

\item{species}{Species for each sound file.}

\item{tpres}{Total duration of presentation in seconds (5 min for sound
transmission)}

\item{intRend.type}{Type of intervals between renditions (constant - mean
natural interval –- median) (depends on species) - read from a database.}

\item{intRend.length}{Length of the interval between rendtions. Only
functional if intRend.type = "custom".}

\item{int.factor}{Factor by which to multiply the length of the interval
between renditions. (0.5, 1, or 2) (for sound transmission experiment).}

\item{nbout}{Number of bouts in presentation (all experiments except sound
transmission – only those in which bout interval type= random)}

\item{bout.length}{Length of each bout (predetermined: 20s – behavior; 10s –
sound transmission; 5s - ZENK)}

\item{intBout.type}{Type of interval between bouts. Can take the values
"random" or "constant".}

\item{intBout.length}{Length of interval between bouts (depending on type of
interval: range, number)}

\item{rate}{Bout rate, in times per minute.}

\item{seed}{Seed for random number generators.}

\item{writewav}{Logical. Should presentation be written into a wave file?}

\item{output}{Name of output file.}

\item{ext}{Extension of output file. Omitted if already specified in output.}

\item{report}{Logical. Should silence gap start times be reported?}

\item{returnoutput}{Logical. Should output filename be returned?}

\item{infoDB}{Databasee containing info on each source file.}

\item{nrend}{Number of renditions per bout.}
}
\description{
Generate auditory presentations based on existing sound files.
}
\details{
If the length of the source audio files is such that the presentation
 would end being longer than T, the last instance of the source sound will be
 truncated in order for the duration of the presentation to be equal to T.
}
\keyword{audio}
\keyword{stimuli}
