% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercalateBlocks.R
\name{intblock}
\alias{intblock}
\title{Intercalate repeats of a block at random or constant intervals}
\usage{
intblock(int.type = "random", int.length = NA, block.length, nblock = NA,
  total, gap_start)
}
\arguments{
\item{int.type}{Type of interval.}

\item{block.length}{Length of the block}

\item{nblock}{Repeats of the block.}

\item{total}{Total length in seconds.}

\item{gap_start}{Logical. Should sequence start with a gap.}
}
\description{
Intercalate repeats of a block at random or constant intervals
}
\details{
A block is considered a time series classfied as signal. A gap is
  silence or noise.

  Intercalates repeat of the blocks, alternating between block and silence. If the total
  duration of the sequence is less than that specified by the used in total,
  a truncated blokc is added at the end until the total length of the
  sequence is equal to that specified in total.
}
\keyword{series}
\keyword{time}
