% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamerize.R
\name{metamerize}
\alias{metamerize}
\title{Metamerize}
\usage{
metamerize(sound, gaps, total, output)
}
\arguments{
\item{sound}{File address of source sound file.}

\item{gaps}{Vector of silence gaps.}

\item{total}{Duration of presentation in seconds.}

\item{output}{Name of output file.}
}
\description{
Construct presentations by repeating an audio signal several times and
introducing silence gaps in between. Requires SoX installed in system.
}
\details{
Will return TRUE ff the function was run succesfully.
}
\keyword{audio,}
\keyword{series}
\keyword{time}
