% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptimMock-data.R
\docType{data}
\name{sptimMock}
\alias{sptimMock}
\title{Mock data of interval renditions typical to a species.}
\format{An object of class \code{data.frame} with 1 rows and 4 columns.}
\usage{
data(sptimMock)
}
\description{
This database is useful to test functions.
}
\keyword{datasets}
