% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTrain.R
\name{genTrain}
\alias{genTrain}
\title{Generate a stimulus train}
\usage{
genTrain(sound)
}
\arguments{
\item{sound}{The sound file to be imported.}
}
\description{
Create auditory stimulus for behavioral experiments.
}
\examples{
genTrain(sound = "#tico#")
}
\keyword{audio}
\keyword{stimuli}
